Revert "feat: scheduled notifications"
