+++ libs/sdk-backend-tiger/src/backend/features/feature.ts
@@ -765,13 +765,6 @@ export function mapFeatures(features: FeaturesMap): Partial<ITigerFeatureFlags> 
             "BOOLEAN",
             FeatureFlagsValues.enableWidgetExportPngImage,
         ),
-        ...loadFeature(
-            features,
-            TigerFeaturesNames.EnableExportToDocumentStorage,
-            "enableExportToDocumentStorage",
-            "BOOLEAN",
-            FeatureFlagsValues.enableExportToDocumentStorage,
-        ),
     };
 }
