+++ libs/sdk-model/api/sdk-model.api.md
@@ -1961,7 +1961,6 @@ export interface IExportOptions {
 // @public
 export interface IExportResult {
     errorMessage?: string;
-    expiresAt?: string;
     exportId: string;
     fileName: string;
     fileUri?: string;
@@ -3422,7 +3421,6 @@ export interface ISettings {
     enableEmbedButtonInKD?: boolean;
     enableExecutionCancelling?: boolean;
     enableExecutionTimestamp?: boolean;
-    enableExportToDocumentStorage?: boolean;
     enableExternalRecipients?: boolean;
     enableGenAIChat?: boolean;
     enableGenAIChatRollout?: boolean;
@@ -4263,7 +4261,6 @@ export interface IVisualizationSwitcherWidgetDefinition extends IVisualizationSw
 
 // @public
 export interface IWebhookAutomationInfo {
-    dashboardTitle?: string;
     dashboardURL: string;
     id: string;
     isCustomDashboardURL?: boolean;
