+++ libs/sdk-ui-dashboard/src/model/store/index.ts
@@ -117,7 +117,6 @@ export {
     selectEnableAlertsEvaluationFrequencySetup,
     selectEnableDateFilterIdentifiers,
     selectEnableSnapshotExportAccessibility,
-    selectEnableExportToDocumentStorage,
 } from "./config/configSelectors.js";
 export type { EntitlementsState } from "./entitlements/entitlementsState.js";
 export {
@@ -475,8 +474,8 @@ export {
 export {
     selectNotificationChannels,
     selectNotificationChannelsCount,
-    selectNotificationChannelsWithoutInPlatform,
-    selectNotificationChannelsCountWithoutInPlatform,
+    selectNotificationChannelsForScheduledExports,
+    selectNotificationChannelsCountForScheduledExports,
 } from "./notificationChannels/notificationChannelsSelectors.js";
 export type { NotificationChannelsState } from "./notificationChannels/notificationChannelsState.js";
 export type { AutomationsState } from "./automations/automationsState.js";
