+++ libs/sdk-ui-ext/api/sdk-ui-ext.api.md
@@ -551,7 +551,6 @@ export interface INotificationsPanelHeaderComponentProps {
 // @public (undocumented)
 export interface INotificationsPanelProps extends INotificationsPanelCustomComponentsProps {
     backend?: IAnalyticalBackend;
-    enableScheduleNotifications?: boolean;
     itemHeight?: number;
     itemPadding?: number;
     itemsGap?: number;
