+++ libs/sdk-ui-ext/src/notificationsPanel/Notification/DefaultNotification.scss
@@ -139,20 +139,6 @@
         align-items: center;
         height: 16px;
         color: var(--gd-palette-complementary-7);
-        & > :first-child {
-            flex-shrink: 1;
-            overflow: hidden;
-            text-overflow: ellipsis;
-            white-space: nowrap;
-        }
-        & > :not(:first-child) {
-            flex-shrink: 0;
-        }
-    }
-
-    &__link {
-        color: var(--gd-palette-complementary-7);
-        text-decoration: underline;
     }
 
     &__error {
