+++ libs/sdk-ui-ext/src/notificationsPanel/data/NotificationsContext.tsx
@@ -1,8 +1,8 @@
 // (C) 2019-2025 GoodData Corporation
-import { IAnalyticalBackend } from "@gooddata/sdk-backend-spi";
-import { UnexpectedSdkError } from "@gooddata/sdk-ui";
 import React from "react";
+import { IAnalyticalBackend } from "@gooddata/sdk-backend-spi";
 import { useNotifications } from "./useNotifications.js";
+import { UnexpectedSdkError } from "@gooddata/sdk-ui";
 
 /**
  * @internal
@@ -32,7 +32,6 @@ export interface INotificationsProviderProps {
     children?: React.ReactNode;
     refreshInterval: number;
     itemsPerPage: number;
-    enableScheduleNotifications: boolean;
 }
 
 /**
@@ -44,15 +43,8 @@ export const NotificationsProvider: React.FC<INotificationsProviderProps> = ({
     workspace,
     refreshInterval,
     itemsPerPage,
-    enableScheduleNotifications,
 }) => {
-    const notifications = useNotifications({
-        backend,
-        workspace,
-        refreshInterval,
-        itemsPerPage,
-        enableScheduleNotifications,
-    });
+    const notifications = useNotifications({ backend, workspace, refreshInterval, itemsPerPage });
 
     return <NotificationsContext.Provider value={notifications}>{children}</NotificationsContext.Provider>;
 };
