+++ libs/sdk-ui-kit/api/sdk-ui-kit.api.md
@@ -1040,8 +1040,6 @@ export interface IBubbleHeaderSeparatorProps {
 // @internal (undocumented)
 export interface IBubbleHoverTriggerProps extends IBubbleTriggerProps {
     // (undocumented)
-    enabled?: boolean;
-    // (undocumented)
     hideDelay?: number;
     // (undocumented)
     hoverHideDelay?: number;
